0324 모듈화 및 pagination 구현

+++ controller/LikeController.js
@@ -1,41 +1,60 @@
-// controller/LikeController.js (수정)
-
+const { ensureAuthorization } = require('../utils/auth');
+const jwt = require('jsonwebtoken');
 const conn = require('../mariadb');
 const { StatusCodes } = require('http-status-codes');
-const { ensureAuthorization } = require('../utils/auth'); // 모듈 import
 
 const addLike = (req, res) => {
     const {bookId} = req.params;
-    const userId = req.user.id; // ensureAuthorization에서 저장한 user 정보 사용
+    
+    let authorization = ensureAuthorization(req, res);
 
-    const sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)";
-    const values = [userId, bookId];
-
-    conn.query(sql, values, (err, results) => {
-        if(err){
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
-
-        return res.status(StatusCodes.OK).json(results);
-    })
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 해주세요."
+        });
+    } else if(authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }else{
+        const sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)";
+        const values = [authorization.id, bookId];
+    
+        conn.query(sql, values, (err, results) => {
+            if(err){
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
 };
 
 const removeLike = (req, res) => {
     const {bookId} = req.params;
-    const userId = req.user.id;
+    
+    let authorization = ensureAuthorization(req, res);
 
-    const sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;`;
-    const values = [userId, bookId];
-
-    conn.query(sql, values, (err, results) => {
-        if(err){
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
-
-        return res.status(StatusCodes.OK).json(results);
-    })
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 해주세요."
+        });
+    } else if(authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }else{
+        const sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;`;
+        const values = [authorization.id, bookId];
+    
+        conn.query(sql, values, (err, results) => {
+            if(err){
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
 };
 
 module.exports = {
